Started on a new database and api
For my fullstack project i do not need all the relationships in my previous api project, hence why i am making this smaller api that is more comprehensible and easier to work with for my next steps. I am doing this to make my "work station" as clean and maintainable as possible.